Full-testing MessageTest Class
